move implementation in separate module and add tests
